Use GLM instead of our own library
Since Math is reduced to a single function, and we have no qualms about including the entirety of GLM in Main, let's just make it a header-only thing and get rid of Math.cpp.
Note that commented out code has not been updated with GLM in mind, so it will no longer work when uncommented.