Include GLM in Math.hpp
It works without it because this is imported in Main after GLM is imported there, but it's still better to also import it here, because it's used here.
Also apparently I forgot to put glm:: before one of the cross calls.
I tested in Main and it works properly now.