Add Mesh code
Our own intermediate format for triangles and meshes, so we can efficiently ray cast and determine which faces should be culled, without having to apply transforms to the original meshes.
Note the check `if (det > -1e-6f) return false;` would normally be `if (det < 1e-6f) return false;`, but is negated here because we actually want to cast from inside the mesh.