Only flip Inside for one of the rays
The even-odd thing is per ray, so only do it for the first ray.
Which ray doesn't matter, because if there are no intersections, all points are on the same side of the mesh.
Therefore, if any point is on the inside, all of them are.
I'm also changing the order of the variables to one that makes more sense to me.
First you determine Intersection, because if that's `true`, the value of Inside is invalid (even if the given point is inside, the others might not be).
Last is Cull, which is determined by both Intersect and Inside.