Precompute edge lengths
Renamed the verts to P# (point, rather than vertex, to avoid confusion with the local variable `v` in RayIntersect), and the edges to E#.
Added L# floats to store edge lengths, which are now calculated in the constructor.
Split the initializer list across a few lines because it was getting too long.

RayIntersect takes the edge length as argument instead of calculating it.
TriangleIntersect sends the appropriate edge length to it.
Not using a reference because a 4 byte float is smaller than an 8 byte pointer.